feat(finance): extract budgets list into feature components (Phase 6b)
- Extract Budgets list logic into BudgetsGrid feature component
- Extract individual budget item JSX into dedicated BudgetCard component
- Enforce keyboard accessibility (role="button", onKeyDown) for budget cards
